Moved manager init related stuff to module.
