+++ vnet/lib/vnet/core/dp_info.rb
@@ -8,8 +8,8 @@
 # Datapath actor's messaging queue for every time we use a manager.
 
 module Vnet::Core
-
   class DpInfo
+    include Vnet::ManagerList
 
     BOOTSTRAP_MANAGER_NAMES = %w(
       host_datapath
@@ -189,46 +189,23 @@ module Vnet::Core
     end
 
     def initialize_bootstrap_managers(timeout, interval = 10.0)
-      bootstrap_managers.tap { |manager_list|
-        manager_list.each { |manager| manager.event_handler_queue_only }
-        manager_list.each { |manager| manager.async.start_initialize }
-        internal_wait_for_initialized(manager_list, timeout, interval).tap { |stuck_managers|
-          next if stuck_managers.nil?
-
-          stuck_managers.each { |manager|
-            Celluloid.logger.warn log_format("#{manager.class.name.to_s.demodulize.underscore} failed to initialize within #{timeout} seconds")
-          }
-          raise Vnet::ManagerInitializationFailed
-        }
-        manager_list.each { |manager| manager.event_handler_active }
-      }
+      initialize_manager_list(bootstrap_managers, timeout, interval)
     end
 
     def initialize_main_managers(datapath_info, timeout, interval = 10.0)
-      main_managers.tap { |manager_list|
-        manager_list.each { |manager| manager.event_handler_queue_only }
-        manager_list.each { |manager| manager.set_datapath_info(datapath_info) }
-        manager_list.each { |manager| manager.async.start_initialize }
-        internal_wait_for_initialized(manager_list, timeout, interval).tap { |stuck_managers|
-          next if stuck_managers.nil?
-
-          stuck_managers.each { |manager|
-            Celluloid.logger.warn log_format("#{manager.class.name.to_s.demodulize.underscore} failed to initialize within #{timeout} seconds")
-          }
-          raise Vnet::ManagerInitializationFailed
-        }
-        manager_list.each { |manager| manager.event_handler_active }
+      initialize_manager_list(main_managers, timeout, interval) { |manager|
+        manager.set_datapath_info(datapath_info)
       }
 
       datapath_manager.async.retrieve(dpid: datapath_info.dpid)
     end
 
     def terminate_bootstrap_managers(timeout = 10.0)
-      internal_terminate_managers(bootstrap_managers, timeout)
+      terminate_manager_list(bootstrap_managers, timeout)
     end
 
     def terminate_main_managers(timeout = 10.0)
-      internal_terminate_managers(main_managers, timeout)
+      terminate_manager_list(main_managers, timeout)
     end
 
     #
@@ -244,42 +221,6 @@ module Vnet::Core
     def internal_initialize_managers(name_list)
       name_list.each { |name|
         instance_variable_set("@#{name}_manager", Vnet::Core.const_get("#{name.to_s.camelize}Manager").new(self))
-      }
-    end
-
-    def internal_wait_for_initialized(manager_list, timeout, interval)
-      manager_list.dup.tap { |waiting_managers|
-        start_timeout = Time.new
-
-        while true
-          # Celluloid.logger.debug log_format('internal_wait_for_initialized interval loop')
-          start_interval = Time.new
-
-          waiting_managers.delete_if { |manager|
-            # Celluloid.logger.debug log_format("internal_wait_for_initialized waiting for #{manager.class.name}")
-            manager.wait_for_initialized(interval - (Time.new - start_interval))
-          }
-
-          return if waiting_managers.empty?
-          return waiting_managers if timeout < (Time.new - start_timeout)
-        end
-      }
-    end
-
-    def internal_terminate_managers(manager_list, timeout)
-      manager_list.each { |manager|
-        begin
-          manager.terminate
-        rescue Celluloid::DeadActorError
-        end
-      }
-
-      start_time = Time.new
-
-      manager_list.each { |manager|
-        next_timeout = timeout - (Time.new - start_time)
-
-        Celluloid::Actor.join(manager, (next_timeout < 0.1) ? 0.1 : next_timeout)
       }
     end
