+++ vnet/lib/vnet/services/vnet_info.rb
@@ -8,6 +8,7 @@
 
 module Vnet::Services
   class VnetInfo
+    include Vnet::ManagerList
 
     SERVICE_MANAGER_NAMES = %w(
       ip_retention_container
@@ -36,23 +37,11 @@ module Vnet::Services
     end
 
     def initialize_service_managers(timeout, interval = 10.0)
-      service_managers.tap { |manager_list|
-        manager_list.each { |manager| manager.event_handler_queue_only }
-        manager_list.each { |manager| manager.async.start_initialize }
-        internal_wait_for_initialized(manager_list, timeout, interval).tap { |stuck_managers|
-          next if stuck_managers.nil?
-
-          stuck_managers.each { |manager|
-            Celluloid.logger.warn log_format("#{manager.class.name.to_s.demodulize.underscore} failed to initialize within #{timeout} seconds")
-          }
-          raise Vnet::ManagerInitializationFailed
-        }
-        manager_list.each { |manager| manager.event_handler_active }
-      }
+      initialize_manager_list(service_managers, timeout, interval)
     end
 
     def terminate_service_managers(timeout = 10.0)
-      internal_terminate_managers(service_managers, timeout)
+      terminate_manager_list(service_managers, timeout)
     end
 
     #
@@ -68,42 +57,6 @@ module Vnet::Services
     def internal_initialize_managers(name_list)
       name_list.each { |name|
         instance_variable_set("@#{name}_manager", Vnet::Services.const_get("#{name.to_s.camelize}Manager").new(self))
-      }
-    end
-
-    def internal_wait_for_initialized(manager_list, timeout, interval)
-      manager_list.dup.tap { |waiting_managers|
-        start_timeout = Time.new
-
-        while true
-          # Celluloid.logger.debug log_format('internal_wait_for_initialized interval loop')
-          start_interval = Time.new
-
-          waiting_managers.delete_if { |manager|
-            # Celluloid.logger.debug log_format("internal_wait_for_initialized waiting for #{manager.class.name}")
-            manager.wait_for_initialized(interval - (Time.new - start_interval))
-          }
-
-          return if waiting_managers.empty?
-          return waiting_managers if timeout < (Time.new - start_timeout)
-        end
-      }
-    end
-
-    def internal_terminate_managers(manager_list, timeout)
-      manager_list.each { |manager|
-        begin
-          manager.terminate
-        rescue Celluloid::DeadActorError
-        end
-      }
-
-      start_time = Time.new
-
-      manager_list.each { |manager|
-        next_timeout = timeout - (Time.new - start_time)
-
-        Celluloid::Actor.join(manager, (next_timeout < 0.1) ? 0.1 : next_timeout)
       }
     end
